ai-services/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
python-docx==1.1.0
PyMuPDF
groq
transformers
torch
sentencepiece
accelerate
faiss-cpu
sentence-transformers
python-pptx
